# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "tiny-expense-tracker-api",
        "version": "0.1.0"
    },
    "paths": {
...
        "/pools": {
            "get": {
                "summary": "Get Pools",
                "operationId": "get_pools_pools_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/StoredMoneyPool"
                                    },
                                    "type": "array",
                                    "title": "Response Get Pools Pools Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Pool",
                "operationId": "create_pool_pools_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MoneyPool"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredMoneyPool"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pools/{pool_id}": {
            "get": {
                "summary": "Get Pool",
                "operationId": "get_pool_pools__pool_id__get",
                "parameters": [
                    {
                        "name": "pool_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Pool Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredMoneyPool"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Modify Pool",
                "operationId": "modify_pool_pools__pool_id__put",
                "parameters": [
                    {
                        "name": "pool_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Pool Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MoneyPoolAttributesUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "summary": "Add Transaction",
                "operationId": "add_transaction_transactions_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transaction"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoredTransaction"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get Transactions",
                "operationId": "get_transactions_transactions_get",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1,
                            "default": 10,
                            "title": "Count"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StoredTransaction"
                                    },
                                    "title": "Response Get Transactions Transactions Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "delete": {
                "summary": "Delete Transaction",
                "operationId": "delete_transaction_transactions__transaction_id__delete",
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Transaction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "summary": "Make Transfer",
                "operationId": "make_transfer_transfer_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferMoneyRequestBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sync-balance/{pool_id}": {
            "post": {
                "summary": "Sync Pool Balance",
                "operationId": "sync_pool_balance_sync_balance__pool_id__post",
                "parameters": [
                    {
                        "name": "pool_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Pool Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncBalanceRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MoneyPool": {
                "properties": {
                    "display_name": {
                        "type": "string",
                        "title": "Display Name"
                    },
                    "balance": {
                        "items": {
                            "$ref": "#/components/schemas/MoneySum-Input"
                        },
                        "type": "array",
                        "title": "Balance"
                    },
                    "is_visible": {
                        "type": "boolean",
                        "title": "Is Visible",
                        "default": true
                    },
                    "last_updated": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Updated"
                    },
                    "display_color": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Color"
                    }
                },
                "type": "object",
                "required": [
                    "display_name",
                    "balance"
                ],
                "title": "MoneyPool"
            },
            "MoneyPoolAttributesUpdate": {
                "properties": {
                    "is_visible": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Visible"
                    },
                    "display_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Name"
                    },
                    "display_color": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Color"
                    }
                },
                "type": "object",
                "title": "MoneyPoolAttributesUpdate"
            },
            "MoneySum-Input": {
                "properties": {
                    "amount": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Amount"
                    },
                    "currency": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "amount",
                    "currency"
                ],
                "title": "MoneySum"
            },
            "MoneySum-Output": {
                "properties": {
                    "amount": {
                        "type": "string",
                        "title": "Amount"
                    },
                    "currency": {
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "amount",
                    "currency"
                ],
                "title": "MoneySum"
            },
            "StoredMoneyPool": {
                "properties": {
                    "display_name": {
                        "type": "string",
                        "title": "Display Name"
                    },
                    "balance": {
                        "items": {
                            "$ref": "#/components/schemas/MoneySum-Output"
                        },
                        "type": "array",
                        "title": "Balance"
                    },
                    "is_visible": {
                        "type": "boolean",
                        "title": "Is Visible",
                        "default": true
                    },
                    "last_updated": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Updated"
                    },
                    "display_color": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Color"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "display_name",
                    "balance",
                    "id"
                ],
                "title": "StoredMoneyPool"
            },
            "StoredTransaction": {
                "properties": {
                    "sum": {
                        "$ref": "#/components/schemas/MoneySum-Output"
                    },
                    "pool_id": {
                        "type": "string",
                        "title": "Pool Id"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "timestamp": {
                        "type": "number",
                        "title": "Timestamp"
                    },
                    "is_diffuse": {
                        "type": "boolean",
                        "title": "Is Diffuse",
                        "default": false
                    },
                    "original_currency": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "sum",
                    "pool_id",
                    "description",
                    "id"
                ],
                "title": "StoredTransaction"
            },
            "SyncBalanceRequestBody": {
                "properties": {
                    "amounts": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "title": "Amounts"
                    }
                },
                "type": "object",
                "required": [
                    "amounts"
                ],
                "title": "SyncBalanceRequestBody"
            },
            "Transaction": {
                "properties": {
                    "sum": {
                        "$ref": "#/components/schemas/MoneySum-Input"
                    },
                    "pool_id": {
                        "type": "string",
                        "title": "Pool Id"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "timestamp": {
                        "type": "number",
                        "title": "Timestamp"
                    },
                    "is_diffuse": {
                        "type": "boolean",
                        "title": "Is Diffuse",
                        "default": false
                    },
                    "original_currency": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "sum",
                    "pool_id",
                    "description"
                ],
                "title": "Transaction"
            },
            "TransferMoneyRequestBody": {
                "properties": {
                    "from_pool": {
                        "type": "string",
                        "title": "From Pool"
                    },
                    "to_pool": {
                        "type": "string",
                        "title": "To Pool"
                    },
                    "sum": {
                        "$ref": "#/components/schemas/MoneySum-Input"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "from_pool",
                    "to_pool",
                    "sum",
                    "description"
                ],
                "title": "TransferMoneyRequestBody"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}